backend/requirements.txt
Flask
Flask-CORS
PyPDF2
python-docx
SpeechRecognition
groq
python-dotenv
speechrecognition
